Remove white background of inline code and modify tag appearance.
The white background on inline code looked ugly and was hard to read
in the table of contents.

The existing tag appearance had a few problems;
the floating block resulted in a bit of empty background
next to the element to the left of the tag span,
and the text colour was pretty unreadable on the background.